pyproject.toml
[build-system]
requires      = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name            = "vgva"
version         = "0.0.1"
description     = "Viral genome variation analysis"
readme          = "README.md"
license         = {text = "GNU General Public License v3 (GPLv3)"}
authors         = [{name = "Bangjun Gong", email = "bangjungong@foxmail.com"}]
classifiers     = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
]
requires-python = ">=3.8"
dependencies    = [
    'numpy>=1.24.4',
    'pandas>=2.0.3',
    'matplotlib>=3.6.2',
    'biopython>=1.83',
    'openpyxl>=3.1.5'
]

[project.urls]
Homepage = "https://github.com/pathogen-detection/vgva"

[tool.setuptools.packages.find]
where = ["src"]
